We do not select or collect any manual or literal RTs now.
Literal: Manually prefix 'RT' before quoting someone and tweeting,
Direct: Someone RT's a tweet but twitter API gives us "RT " in the text

Why?
1. No meaning in RT-ing RTs. They don't help trends either.
2. Twitter API truncates the text somehow when it collects such tweets
adding a special unicode character for ellipse(...) which confuses our
template engine and makes a tweet look ugly and especially unclickable
when it is a link.